Add initial logging and host configuration
